Extract common Admin handlers.
Both the Finatra and Movies examples duplicated the Admin servlet. And they both had put the admin API on the public HTTP port. The Finagle and Unfiltered example lacked any admin API.

This changes the admin servlet to several Finagle ones and installs them on the admin port of Twitter Server. It does this for the Finagle and Unfiltered examples. TODOs detail what needs to happen with the Finagle and Movies examples.

closes #57