Migrate Movies example from Finatra to Finagle.
Finatra is built against Twitter Server 6.17, and Treode's utilities for Twitter Server are built against 6.22. An incompatibility between the two prevented a gradual migration from Finatra to Finagle, so all the resources needed to be changed at once. Furthermore, Finatra provides MockApp for testing, but with Finagle one must setup a server and use a client for testing, so all the tests migrated from MockApp to RestAssured in this same CL.