Add a parameter for batch size to scan.
The changes over the past couple weeks now permit scans run in batches (or chunks) all the way from the replica up thru the stack to the HTTP client. That effort has paid off. Downloading the list of movies and actors runs 5x faster. Downloading the list of roles is now possible.