Combine twitter-util with twitter-server; parse addresses consistently.
These were split because one used finagle and the other didn't. However, StoreKit should parse internet addresses the same way HttpServer does, and the utilities for that are in finagle. Then both twitter-util and twitter-server have the same dependencies. Then having them separated just creates unnecessary complications.

This changes command line arguments. What was once

    -peerPort=1234

is now

    -peerAddr=:1234

The advancement introduced here is that, first, one can bind on a non-default address:

    -peerAddr=129.168.1.7:1234

And second, one can specify which address to share with peers:

    -sharePeerAddr=129.168.1.7:1234

as a way to override the automatic mechanism, which might compute something like:

    -sharePeerAddr=localhost:1234