Remove engaged state in Dispatcher.
The dispatcher had a notion of being "engaged." It would deliver a batch of messages to a first receiver, and it would hold all waiting receivers until the first returned the rejects. That was unnecessary. After delivering a batch of messages to a first receiver, it could deliver another batch to a second receiver. It needed to allow, but not require, the second coordination with the first receiver (that is, replacing rejects).

https://forum.treode.com/t/dispatcher-can-dump-notion-of-engaged/54

closes #105