Fence stub globals into special object.
When the compiler needed an implicit StubScheduler, and it could not find one in scope, it happily grabbed the global one. It would have saved us much debugging time to get an error instead. This hides the globals in a differently named class so that one must make them available explicitly.