Simplify API for AsyncQueue and add ScalaDoc.
This API is much simpler to explain, and yet it can mostly support the old API. This CL makes minor changes the current users of AsyncQueue. The AtomicMover and PaxosMover now use only the simpler API. The DiskDrives class continues to use the old confusing methods. That will be replaced soon, and then a later CL will remove the old methods.