Remove fiber from Dispatcher, and add setter for batch.
The fiber wraps an ArrayDeque with synchronized to serialize tasks. The Dispatcher can more efficiently wrap ArrayDeque and UnrolledBuffer with synchronized to serialized tasks, rather than using a fiber for that.

A setter for batch allows the batch number to be raised after creation. This will help cleanup circular dependencies in creating components during log replay. Specifically, log replay needs a dispatcher, but the value of the batch number is not known until log replay completes.