Clean up API for PicklerRegistry.
The API had some methods that checked the amount of data consumed, and others that did not. It had some methods that discarded consumed bytes from a PagedBuffer and others that did not. This was confusing. Now all methods check the amount of data consumed, and all methods let the caller decide how to discard consumed bytes.

This change closes [this bug](https://forum.treode.com/t/add-unpickle-method-to-picklerregistry-that-skips-checking-consumption/43), however it does so much differently than originally proposed.