Create SegmentAllocator and use it for logging.
Create the SegmentAllocator to use compressed bitmap for tracking free segments. Adapt the LogWriter to allocate a segment, and to write only the records that fit into the remainder of the segment. Adapt the LogReader to jump from the end of one segment to the beginning of the next. This effort included tracking the log head from checkpoint to checkpoint.