Add detacher to mediate draining a drive.
This is the new shiny version of what was called the Multiplexer in the old package. That is, this serves the same purpose as that: ensure that a disk drive is detached after completing a flush, if a flush is in progress. The Detacher is more appropriately named, better documented, and far simpler.