Introduce bootstrap classes.
The SegmentLedger depends on the Disk (to log records and write itself); the Disk depends on the SegmentLedger (to track page allocations). As a result, many classes had a field for the SegmentLedger that was initially null and set later. The BootstrapDisk allows the SegmentLedger to startup, and then be injected into the constructors of the other classes. Later the BootstrapDisk is replaced by the live Disk in the SegmentLedger only.

The user's checkpoints and compactors need the Disk (to write pages); the Disk needs the checkpoints and compactors (to free log segments and page segments). As a result, the Checkpointer and Compactor were null in many classes and set later. The refactoring for the BootstrapDisk created a nice opportunity to clean that up as well. Now the Checkpointer and Compactor are injected into the other classes. Later the user checkpoints and compactors are set in those two locations only.