Have ResourceHandler combine schema and store.
An intermediate class was combining these two things, but that created two issues: the intermediary could not know the proper mechanism to report errors (exceptions vs status codes), and it complicated attaching the schema to the catalogs. This change hands all the responsibility to the ResourceHandler to prepare for later changes involving errors and catalogs.