Check that Disk system grows files properly.
A StubFile used a byte array directly, and the StubFileSystem mapped to byte arrays. However this meant that when the StubFile grew the array, the StubFileSystem will not be updated. To ensure the StubFileSystem tracks the growing arrays, this change introduces StubFile.Data. That holds the growable byte array, StubFile works through that, and the StubFileSystem maps paths to that.