Support for range scan and pagination.
The HTTP server that receives the user's request sends the start point to the replicas when scanning, and the replicas return some batch of data unaware of the user's pagination limit and end key. The server then filters the data to find the limit and end. It would be good, in a future change, to convey the parameters all the way to the replicas.